Phase_2(annuity,loan_principal,payment_amount)
In this stage, we will add new behavior to the calculator:

First, you will ask the user which parameter they want to calculate. The calculator should be able to calculate the number of monthly payments, the monthly payment amount, and the loan principal.
Then, we need to ask them to input the remaining values.
Finally, we will compute and output the value that they wanted.